Move AJ adapter
