+++ lib/active_job/queue_adapters/resque_adapter.rb
